Level Class (mainly)
I did a bunch of stuff.

-First off, I added the level class, and moved madisons code over to
that, which works through that now. I didnt delete the method in game1
just in case, but it now goes through the level class to draw that grid
of char's.

-I implemented the sprite dictionary, so make sure when you add new
textures you add them to that and give them appropriate keys.

-I added some placeholder sprites for walls and floor tiles

-The setupLevel method in the level class doesnt currently work. I
explained why in the method description in the code. I cant do anything
else until wes and joel do some more architecture work.

-I also made it so it draws the grid of chars before the player because
that was bugging me.