Enemies, Level, ImmobilesManager
1) Added basic movement to the enemies. If they haven't detected you,
they will walk in a square. If they have, they will follow you using the
shortest path. I also put in a wait time between updates for them, so
that they're slower than the enemy
2) I made an ImmobilesManager that contains all of the Immobiles and
through which you can update and draw them. I think we should maybe get
rid of this in favor of an overall GameObjectsManager.
3) Implemented the remaining code for reading in walls and floor in
Level, and made it so using SetupLevel doesn't result in an error