Beginning of shaders
Dims the screen and creates a light source around the player. Due to the
nature of the shader anything completely black will NOT be lit because
it's alpha is zero and the shader works on multiplication. I also have
an echo effect in there but I might not use it