Its a game now!
Made some balances so we can actually show it off as a game. I also
switched the level editor that was in there to the one that is the 25x15
array that i used to make the new levels. Its not necessary but its
convenient since the levels can't be bigger than that now anyways