Adam's Update
So I added some of the new enemies we talked about. Enemy is now there
for basically organizational purposes, and what was the enemy is now
BatEnemy. I also added a sleeping bat enemy, which does work entirely,
as well as a spider enemy. The spider enemy almost works, but i just
have to figure out a couple more things. I also added a bunch of stuff
to the level editor, and just included the actual code this time instead
of a release version. If you want you can go deep into the files and
find some of the new levels ive made, i havent implemented any of them
yet though because its not done for one, and its almost definitely gonna
change when we finish implementing everything we want to. So yeah, i
think that's everything